requirements.txt
#requirements.txt

Django>=5.0,<6.0
mysqlclient>=2.1,<3.0
python-decouple>=3.8,<4.0
python-dateutil>=2.8,<3.0
django-widget-tweaks==1.4.12